light-sensor working code

+++ src/light-sensor/src/Opt3001Base.h
@@ -6,7 +6,8 @@
 class Opt3001Base {
 public:
 	virtual bool init() { return true; }
-	virtual float readLux()  { return 1.0f; }
+	virtual uint16_t readLux()  { return 1; }
+	// virtual float readLux()  { return 1.0f; }
 };
 
 #endif
